+++ tests/testthat/test-skim.R
@@ -667,6 +667,94 @@ test_that("skim returns expected response for lubridate Timespan vectors", {
   expect_identical(input$Timespan.n_unique, 1L)
 })
 
+test_that("skim handles objects containing haven_labelled vectors: double", {
+  dt <- tibble::tibble(x = haven::labelled(c(1, 2, 3), c(A = 1, B = 2, C = 3)))
+  input <- skimr::skim(dt)
+
+  # dimensions
+  expect_n_rows(input, 1)
+  expect_n_columns(input, 12)
+
+  # classes
+  expect_s3_class(input, "skim_df")
+  expect_s3_class(input, "tbl_df")
+  expect_s3_class(input, "tbl")
+  expect_s3_class(input, "data.frame")
+  expect_named(input, c(
+    "skim_type", "skim_variable", "n_missing", "complete_rate",
+    "numeric.mean", "numeric.sd", "numeric.p0", "numeric.p25",
+    "numeric.p50", "numeric.p75", "numeric.p100", "numeric.hist"
+  ))
+
+  # attributes
+  attrs <- attributes(input)
+  expect_equal(attrs$data_rows, 3)
+  expect_equal(attrs$data_cols, 1)
+  expect_equal(attrs$df_name, "`dt`")
+  expect_equal(
+    attrs$skimmers_used,
+    list(
+      numeric = c("mean", "sd", "p0", "p25", "p50", "p75", "p100", "hist")
+    )
+  )
+
+  # values
+  expect_identical(input$skim_type, "numeric")
+  expect_identical(input$skim_variable, "x")
+  expect_identical(input$n_missing, 0L)
+  expect_equal(input$complete_rate, 1)
+  expect_equal(input$numeric.mean, 2, tolerance = 0.1)
+  expect_equal(input$numeric.sd, 1, tolerance = 0.1)
+  expect_equal(input$numeric.p0, 1, tolerance = 0.1)
+  expect_equal(input$numeric.p25, 1.5, tolerance = 0.1)
+  expect_equal(input$numeric.p50, 2, tolerance = 0.1)
+  expect_equal(input$numeric.p75, 2.5, tolerance = 0.1)
+  expect_equal(input$numeric.p100, 3, tolerance = 0.1)
+})
+
+test_that("skim handles objects containing haven_labelled vectors: character", {
+  dt <- tibble::tibble(x = haven::labelled(
+    c("a", "b", "c"),
+    c(A = "a", B = "b", C = "c")
+  ))
+  input <- skimr::skim(dt)
+
+  # dimensions
+  expect_n_rows(input, 1)
+  expect_n_columns(input, 9)
+
+  # classes
+  expect_s3_class(input, "skim_df")
+  expect_s3_class(input, "tbl_df")
+  expect_s3_class(input, "tbl")
+  expect_s3_class(input, "data.frame")
+  expect_named(input, c(
+    "skim_type", "skim_variable", "n_missing", "complete_rate",
+    "character.min", "character.max", "character.empty",
+    "character.n_unique", "character.whitespace"
+  ))
+
+  # attributes
+  attrs <- attributes(input)
+  expect_equal(attrs$data_rows, 3)
+  expect_equal(attrs$data_cols, 1)
+  expect_equal(attrs$df_name, "`dt`")
+  expect_equal(
+    attrs$skimmers_used,
+    list(character = c("min", "max", "empty", "n_unique", "whitespace"))
+  )
+
+  # values
+  expect_identical(input$skim_variable, "x")
+  expect_identical(input$skim_type, "character")
+  expect_identical(input$n_missing, 0L)
+  expect_equal(input$complete_rate, 1., tolerance = .001)
+  expect_identical(input$character.min, 1L)
+  expect_identical(input$character.max, 1L)
+  expect_identical(input$character.empty, 0L)
+  expect_identical(input$character.n_unique, 3L)
+  expect_identical(input$character.whitespace, 0L)
+})
 
 test_that("skim handles objects with multiple classes", {
   dat <- seq(as.Date("2011-07-01"), by = 1, len = 10)