+++ NEWS.md
@@ -3,6 +3,10 @@
 ### NEW FEATURES
 
 *   skim() used within a function now prints the data frame name.
+*   we have improved the interaction between focus() and the print methods.
+
+    *  columns selected in focus() are shown in the correct order
+    *  some edge cases relating to empty skim types have been improved
 
 # skimr 2.1.3