+++ NEWS.md
@@ -4,9 +4,9 @@
 
 *   skim() used within a function now prints the data frame name.
 *   we have improved the interaction between focus() and the print methods.
-
     *  columns selected in focus() are shown in the correct order
     *  some edge cases relating to empty skim types have been improved
+*   we have improved performance when handling large data with many columns.
 
 # skimr 2.1.3