+++ NEWS.md
@@ -7,6 +7,7 @@
 
     *  columns selected in focus() are shown in the correct order
     *  some edge cases relating to empty skim types have been improved
+*   we have improved performance when handling large data with many columns.
 
 # skimr 2.1.3