+++ NEWS.md
@@ -4,7 +4,6 @@
 
 *   skim() used within a function now prints the data frame name.
 *   we have improved the interaction between focus() and the print methods.
-
     *  columns selected in focus() are shown in the correct order
     *  some edge cases relating to empty skim types have been improved
 *   we have improved performance when handling large data with many columns.
@@ -12,7 +11,10 @@
 ### Minor improvements
 
 *   Replace the Suppporting Additional Objects vignette
-with Extending skimr.  Remove sf from Suggests.
+    with Extending skimr.  Remove sf from Suggests.
+*   Default support for `haven_labelled` columns is now supported. These
+    columns are summarized using skimmers for the underlying data, typically
+    either numeric or character.
 
 # skimr 2.1.3